+++ DesignToolbox/DesignToolboxUITests/Screenshots/Documentation Screenshots - Controls.swift
@@ -296,6 +296,34 @@ final class 📄_ControlsDocumentationScreenshots: AppTestCase {
                        ACDC.chipPickerHeight,
                        app)
     }
+
+    // MARK: - Text inputs
+
+    @MainActor
+    func testMakeScreenshotsForDocumentation_TextInput() {
+        let app = launchApp()
+        goToComponentsSheet(app)
+        swipeFromDownToUp(app)
+        swipeFromDownToUp(app)
+        swipeFromDownToUp(app)
+        swipeFromDownToUp(app)
+        waitForButtonToAppear(withWording: "app_components_textInput_label", app)
+        tapButton(withWording: "app_components_textInput_label", app)
+
+        wait(2)
+
+        tapOtherElement(withA11yIdentifier: A11YIdentifiers.configurationSwitchOutlined, app)
+        tapOtherElement(withA11yIdentifier: A11YIdentifiers.configurationSwitchLeadingIcon, app)
+
+        wait(2)
+
+        takeScreenshot(named: "component_textInput_",
+                       ACDC.textInputX,
+                       ACDC.textInputY,
+                       ACDC.textInputWidth,
+                       ACDC.textInputHeight,
+                       app)
+    }
 }
 
 // swiftlint:enable required_deinit