+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- [Library] TextInput component (Orange-OpenSource/ouds-ios#406)
+- [Library] `Text input` component (Orange-OpenSource/ouds-ios#406)
 
 ### Changed